requirements.txt
PyQt6>=6.6.0
websockets>=12.0
cryptography>=41.0.0
pyinstaller>=6.0.0
Pillow>=10.0.0
